backend/requirements.txt
fastapi
uvicorn
kubernetes>=26.1.0
python-jose
passlib[bcrypt]
